// generated/src/app/api-models/member-other-coverage.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MemberOtherCoverage {
  seqMembOthCov: number;
  seqMembId: number;
  seqSubsId: number;
  carrierCode: string;
  coverageEffecDate: Date;
  termDate: Date;
  termReason: string;
  policyNumber: string;
  cobCode: string;
  userDefined1: string;
  userDefined2: string;
  otherFamCoverage: string;
  personNumber1: string;
  personNumber2: string;
  personNumber3: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  entitlementCode: string;
  verifOthCarrierDate: Date;
  verifStatusCode: string;
  verifSourceLastName: string;
  verifSourceFirstName: string;
  verifSourcePhoneNo: string;
  comments: string;
  coverageTypeCode: string;
  employerGroup: string;
  planTypeCode: string;
  productType: string;
  othCarrierSubsId: string;
  othCarrierSubsLastName: string;
  othCarrierSubsFirstName: string;
  othCarrierSubsDob: Date;
  othCarrierSubsGender: string;
  relationshipCode: string;
  othCarrierCobFlag: string;
  depDetermRuleCode: string;
  secDepDetermRuleCode: string;
  paymentPolicyCode: string;
  preCertReqdFlag: string;
  caseNumber: string;
  diagnosisCode: string;
  statusCode: string;
  facility: string;
  eventDate1: Date;
  eventDate1TypeCode: string;
  eventDate2: Date;
  eventDate2TypeCode: string;
  eventDate3: Date;
  eventDate3TypeCode: string;
  userDefined3: string;
  userDefined4: string;
  userDefined5: string;
  workingAgedStatus: string;
  cobhsUserDate1: Date;
  cobhsUserDate2: Date;
}